feat: cơ chế tiền tip
Tip chance và TipMultiplier đang fixed cứng

+++ Assets/_Project/Modules/Restaurant/Scripts/Domain/Models/Order.cs
@@ -11,10 +11,10 @@ namespace LabDiner.Restaurant.Model
     {
         public Dictionary<CoreStation, int> OrderDict; // CoreStation là món ăn, int là số lượng
         public GuestContext OrderBy;
-        public float Profit = 0; // Lợi nhuận thu được từ order này, sẽ được câph nhật lại sau
+        public double Profit = 0; // Lợi nhuận thu được từ order này, sẽ được câph nhật lại sau
         public bool IsServed = false; // Đánh dấu xem order đã được phục vụ hay chưa
 
-        public Order(Dictionary<CoreStation, int> orderDict, GuestContext orderBy, float profit, bool isServed)
+        public Order(Dictionary<CoreStation, int> orderDict, GuestContext orderBy, double profit, bool isServed)
         {
             OrderDict = orderDict;
             OrderBy = orderBy;
